Enable Frontend & Postadressen-svc to run as non-root containers
In environments where restricted pod security policies are enabled
the bitnami/nginx container (used by the frontend) will fail to start.
This is a known issue explained in
https://github.com/bitnami/bitnami-docker-nginx/issues/121.
The mentioned work-around in the nginx.conf has been implemented.

To enable Postadressen-svc to run as non-root, a change in the
concerning Dockerfile has been made. When running this container
as non-root in Kubernetes, this should be done with 'runAsUser: 1001'.